+++ Sources/Sentry/SentrySessionTracker.m
@@ -5,7 +5,7 @@
 #import "SentryHub.h"
 #import "SentryInternalNotificationNames.h"
 #import "SentryLog.h"
-#import "SentrySDK.h"
+#import "SentrySDK+Private.h"
 
 #if SENTRY_HAS_UIKIT
 #    import <UIKit/UIKit.h>