+++ Tests/SentryTests/SentrySDK+Tests.h
@@ -1,0 +1,11 @@
+#import <Sentry/Sentry.h>
+
+NS_ASSUME_NONNULL_BEGIN
+
+@interface SentrySDK (Tests)
+
++ (void)setCurrentHub:(SentryHub *)hub;
+
+@end
+
+NS_ASSUME_NONNULL_END